requirements.txt
# Everything needed in production

# Flask
Flask>=1.1.1
MarkupSafe>=1.0
Werkzeug>=0.14.1
Jinja2>=2.10
itsdangerous>=1.1.0
click>=5.0

# Database
Flask-SQLAlchemy>=2.3.2
Flask-Marshmallow>=0.12.0
psycopg2>=2.7.5
SQLAlchemy>=1.2.11
PyMySQL>=0.9.2

# Deployment
gunicorn>=19.1.1

# Auth
Flask-Login>=0.4.1
Flask-Bcrypt>=0.7.1

# Caching
Flask-Caching>=1.0.0

# Debug toolbar
Flask-DebugToolbar>=0.10.1

# Environment variable parsing
environs>=4.0.0

# Restful
flask-restful>=0.3.6
marshmallow-sqlalchemy>=0.15.0
flask-cors>=3.0.7

#other
arrow>=0.14.4
apispec[yaml]
apispec-webframeworks


# From pubCode
certifi>=2019.3.9
chardet>=3.0.4
DBUtils>=1.3
idna>=2.8
pika>=1.0.1
redis>=3.2.1
requests>=2.22.0
urllib3>=1.25.3
Flask>=1.1.1
flask_buzz>=0.1.14